push

+++ src/main/java/org/wlgzs/xf_mall/service/impl/ProductServiceImpl.java
@@ -137,19 +137,25 @@ public class ProductServiceImpl implements ProductService {
 
     //删除商品
     @Override
-    public void delete(long productId, HttpServletRequest request) {
+    public String delete(long productId, HttpServletRequest request) {
         Product product = productRepository.findById(productId);
-        String path = request.getSession().getServletContext().getRealPath("/");
-        String image = product.getProduct_picture();
+        if(product != null){
+            String path = request.getSession().getServletContext().getRealPath("/");
+            String image = product.getProduct_picture();
 
-        String[] img = image.split(",");
-        for (int i = 0; i < img.length; i++) {
-            File file = new File(path+""+img[i].substring(1,img[0].length()));
-            if (file.exists() && file.isFile()) {
-                file.delete();
+            String[] img = image.split(",");
+            for (int i = 0; i < img.length; i++) {
+                File file = new File(path+""+img[i].substring(1,img[0].length()));
+                if (file.exists() && file.isFile()) {
+                    file.delete();
+                }
             }
+            productRepository.deleteById(productId);
+            return "成功";
+        }else {
+            return "失败";
         }
-        productRepository.deleteById(productId);
+
     }
 
     //通过id查找商品
@@ -197,50 +203,55 @@ public class ProductServiceImpl implements ProductService {
 
     //修改商品
     @Override
-    public void edit(long productId, String product_details, MultipartFile[] myFileNames, HttpSession session,
+    public String edit(long productId, String product_details, MultipartFile[] myFileNames, HttpSession session,
                      HttpServletRequest request) {
-        String realName = "";
-        String[] str = new String[myFileNames.length];
-        for (int i = 0; i < myFileNames.length; i++) {
-            if (!myFileNames[i].getOriginalFilename().equals("")) {
-                String fileName = myFileNames[i].getOriginalFilename();
-                String fileNameExtension = fileName.substring(fileName.indexOf("."), fileName.length());
+        Product product = productRepository.findById(productId);
+        if(product != null){
+            String realName = "";
+            String[] str = new String[myFileNames.length];
+            for (int i = 0; i < myFileNames.length; i++) {
+                if (!myFileNames[i].getOriginalFilename().equals("")) {
+                    String fileName = myFileNames[i].getOriginalFilename();
+                    String fileNameExtension = fileName.substring(fileName.indexOf("."), fileName.length());
 
-                // 生成实际存储的真实文件名
-                realName = UUID.randomUUID().toString() + fileNameExtension;
-                // "/upload"是你自己定义的上传目录
-                String realPath = session.getServletContext().getRealPath("/upload");
-                File uploadFile = new File(realPath, realName);
-                System.out.println(uploadFile);
-                try {
-                    myFileNames[i].transferTo(uploadFile);
-                } catch (IOException e) {
-                    e.printStackTrace();
+                    // 生成实际存储的真实文件名
+                    realName = UUID.randomUUID().toString() + fileNameExtension;
+                    // "/upload"是你自己定义的上传目录
+                    String realPath = session.getServletContext().getRealPath("/upload");
+                    File uploadFile = new File(realPath, realName);
+                    System.out.println(uploadFile);
+                    try {
+                        myFileNames[i].transferTo(uploadFile);
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+                if (!myFileNames[i].getOriginalFilename().equals("")) {
+                    str[i] = request.getContextPath() + "/upload/" + realName;
                 }
             }
-            if (!myFileNames[i].getOriginalFilename().equals("")) {
-                str[i] = request.getContextPath() + "/upload/" + realName;
+            StringBuffer stringBuffer = new StringBuffer();
+            for (int i = 0; i < str.length; i++) {
+                if (!myFileNames[i].getOriginalFilename().equals("")) {
+                    stringBuffer.append(str[i] + ",");
+                }
             }
-        }
-        StringBuffer stringBuffer = new StringBuffer();
-        for (int i = 0; i < str.length; i++) {
-            if (!myFileNames[i].getOriginalFilename().equals("")) {
-                stringBuffer.append(str[i] + ",");
+            String product_picture = stringBuffer.substring(0, stringBuffer.length() - 1);
+            Map<String, String[]> properties = request.getParameterMap();
+            try {
+                BeanUtils.populate(product, properties);
+            } catch (IllegalAccessException e) {
+                e.printStackTrace();
+            } catch (InvocationTargetException e) {
+                e.printStackTrace();
             }
+            product.setProduct_details(product_details);
+            product.setProduct_picture(product_picture);
+            productRepository.save(product);
+            return "修改成功";
+        }else{
+            return "修改失败";
         }
-        String product_picture = stringBuffer.substring(0, stringBuffer.length() - 1);
-        Product product = productRepository.findById(productId);
-        Map<String, String[]> properties = request.getParameterMap();
-        try {
-            BeanUtils.populate(product, properties);
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        } catch (InvocationTargetException e) {
-            e.printStackTrace();
-        }
-        product.setProduct_details(product_details);
-        product.setProduct_picture(product_picture);
-        productRepository.save(product);
     }
 
     //遍历所有分类  搜索分类
@@ -338,19 +349,23 @@ public class ProductServiceImpl implements ProductService {
 
     //删除分类
     @Override
-    public void deleteCategory(long categoryId) {
+    public String deleteCategory(long categoryId) {
         ProductCategory productCategory = productCategoryRepository.findById(categoryId);
-        if(productCategory.getParent_name().equals("0")){
-            List<ProductCategory> productCategories = productCategoryRepository.findByCategoryParentName(productCategory.getCategory_name());
-            if(productCategories.size()!=0){
-                long[] Ids = new long[productCategories.size()];
-                for (int i = 0; i < Ids.length; i++) {
-                    Ids[i] = productCategories.get(i).getCategoryId();
+        if(productCategory != null){
+            if(productCategory.getParent_name().equals("0")){
+                List<ProductCategory> productCategories = productCategoryRepository.findByCategoryParentName(productCategory.getCategory_name());
+                if(productCategories.size()!=0){
+                    long[] Ids = new long[productCategories.size()];
+                    for (int i = 0; i < Ids.length; i++) {
+                        Ids[i] = productCategories.get(i).getCategoryId();
+                    }
+                    productCategoryRepository.deleteByIds(Ids);
                 }
-                productCategoryRepository.deleteByIds(Ids);
             }
+            productCategoryRepository.deleteById(categoryId);
+            return "删除成功";
         }
-        productCategoryRepository.deleteById(categoryId);
+        return "删除失败";
     }
 
     //修改分类
